feat(engine): per-request transcript-cache bypass (DecodeParams::bypass_transcript_cache)
bd-rt-cache-bypass-xqz1: the live listen driver re-decodes a rolling
buffer every step — the per-model result LRU can never hit there, yet
each miss copies up to 16 MiB of samples into the cache and evicts
useful batch entries. New field bypass_transcript_cache (default false;
stored inverted so derive(Default) stays correct) skips lookup AND
insert before fingerprint construction (no keyspace pollution);
FW_TRANSCRIPT_CACHE=0 remains the process-wide operator kill-switch
(effective caching = env && !bypass). transcribe_samples_batch honors
it per job for free: params key the coalescing groups, so bypass jobs
never coalesce with cached jobs (asserted). Tracing debug line on
bypass for harness trace clarity.

Tests: default-off + cache-key-inequality assertions; model-gated e2e
proving bypass runs keep the LRU empty while doing real decode work
(sampled_tokens > 0 both runs) and the no-bypass control is a pure
cache hit (sampled_tokens == 0). Passed 2/2 with tiny.en + jfk.

Also in this commit (main-repair, other agents' in-flight state):
complete the earshot verdict edit in Cargo.toml per its own comment
(dep line restored for the ignored re-evaluation battery the comment
says is kept), and clear the battery's same-type-cast clippy errors.
Gates: fmt clean, clippy --all-targets -D warnings exit 0.

+++ src/listen.rs
@@ -1206,7 +1206,7 @@ mod tests {
 
         let music = music_signal(SAMPLE_RATE * 6, -14.0);
         let noise = lcg_noise(SAMPLE_RATE * 4, 0xC0FFEE, -20.0);
-        let hum = hum_signal(SAMPLE_RATE as usize * 4, -18.0);
+        let hum = hum_signal(SAMPLE_RATE * 4, -18.0);
         let mut violations: Vec<String> = Vec::new();
         let row = |name: &str, samples: &[f32], scores: &[f32], violations: &mut Vec<String>| {
             let voiced = scores.iter().filter(|&&s| s >= 0.5).count();
@@ -1287,7 +1287,7 @@ mod tests {
     #[test]
     fn earshot_eval_cost_per_frame_is_far_under_step_budget() {
         // may cost at most ~1 ms per step => <=200 us per 16 ms predict.
-        let audio = music_signal(SAMPLE_RATE as usize * 10, -14.0);
+        let audio = music_signal(SAMPLE_RATE * 10, -14.0);
         let mut det = earshot::Detector::default_boxed();
         let frames: Vec<&[f32]> = audio
             .as_chunks::<EARSHOT_FRAME_SAMPLES>()
